Mailmon v2.1
Updated to use smtplib instead of relying on the local MTA to send the
emails. Also fixed some minor bugs here and there. Take a look at the
new SMTP-related options that need to be set in the beginning of the
file.

Next up on the TODO list is to remove the requirement of knowing python
and allow commands to be in the form of bash scripts or other executable
files.